requirements.txt
requests>=2.31.0
openai>=1.12.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0

# dev / testing
pytest>=7.0.0
